+++ diesel_cli/tests/print_schema/print_schema_type_renaming/postgres/schema.sql
@@ -1,0 +1,7 @@
+CREATE TYPE user_job AS ENUM ('programmer', 'director', 'writer', 'mathematician');
+
+CREATE TABLE users (
+    id INTEGER PRIMARY KEY,
+    job user_job NOT NULL
+);
+